+++ CHANGELOG.md
@@ -66,3 +66,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Security: Explicitly permitted `Symbol` and `DateTime` classes in YAML loading.
 
 [0.0.1]: https://github.com/joshkovach/astro-subframe-organizer/releases/tag/v0.0.1
+[0.0.2]: https://github.com/joshkovach/astro-subframe-organizer/compare/v0.0.1...v0.0.2